Add api packages on repository and service packages and move interfaces to them
