--- src/main/java/com/mercadolibre/be_java_hisp_w31_g04/repository/IProductRepository.java
@@ -1,4 +1,0 @@
-package com.mercadolibre.be_java_hisp_w31_g04.repository;
-
-public interface IProductRepository {
-}
