+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/repository/ProductRepositoryImpl.java
@@ -1,5 +1,6 @@
 package com.mercadolibre.be_java_hisp_w31_g04.repository;
 
+import com.mercadolibre.be_java_hisp_w31_g04.repository.api.IProductRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
