+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/repository/UserRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.mercadolibre.be_java_hisp_w31_g04.repository;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mercadolibre.be_java_hisp_w31_g04.model.User;
-import lombok.Getter;
+import com.mercadolibre.be_java_hisp_w31_g04.repository.api.IUserRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.ResourceUtils;
 
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class UserRepositoryImpl implements IUserRepository{
+public class UserRepositoryImpl implements IUserRepository {
     private List<User> listOfUsers= new ArrayList<>();
 
     public UserRepositoryImpl() throws IOException{
