+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/repository/api/IProductRepository.java
@@ -1,0 +1,4 @@
+package com.mercadolibre.be_java_hisp_w31_g04.repository.api;
+
+public interface IProductRepository {
+}
