+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/repository/api/IUserRepository.java
@@ -1,4 +1,4 @@
-package com.mercadolibre.be_java_hisp_w31_g04.repository;
+package com.mercadolibre.be_java_hisp_w31_g04.repository.api;
 
 import com.mercadolibre.be_java_hisp_w31_g04.model.User;
