+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/service/ProductServiceImpl.java
@@ -1,7 +1,8 @@
 package com.mercadolibre.be_java_hisp_w31_g04.service;
 
+import com.mercadolibre.be_java_hisp_w31_g04.service.api.IProductService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ProductServiceImpl implements IProductService{
+public class ProductServiceImpl implements IProductService {
 }
