+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g04/service/api/IUserService.java
@@ -1,4 +1,4 @@
-package com.mercadolibre.be_java_hisp_w31_g04.service;
+package com.mercadolibre.be_java_hisp_w31_g04.service.api;
 
 import com.mercadolibre.be_java_hisp_w31_g04.dto.FollowersCountDto;
 import com.mercadolibre.be_java_hisp_w31_g04.dto.UserFollowedDto;
